feat: refactor project components and enhance content management
- Replaced ProjectList with ContentList for improved content display.
- Introduced new components: ContentList, LinkListItem, ProjectListItem, and SeparatorListItem.
- Updated project structure by removing obsolete components and integrating new styles.
- Enhanced GraphQL queries to support new content types and improve data fetching.
- Added remote image patterns to Next.js configuration for better image handling.
