+++ packages/web/src/app/[projectSlug]/page.tsx
@@ -1,7 +1,7 @@
 import NextLink from "next/link";
 import * as React from "react";
 
-import { ProjectList } from "@/components/project-list";
+import { ContentList } from "@/components/content-list";
 import { getHomePageData } from "@/lib";
 import { getCollectionItems, getFirstItem } from "@/lib/graphql";
 import layoutStyles from "@/styles/layout.module.css";
@@ -18,7 +18,7 @@ export default async function ProjectPage() {
           <NextLink href="/">ALEXANDRA LEYKAUF</NextLink>
         </h1>
       </header>
-      <ProjectList items={contentItems} />
+      <ContentList items={contentItems} />
       <footer className={layoutStyles.footer}>
         <NextLink
           href="https://www.ourcompany.ch/"
