+++ packages/web/src/components/content-list/index.ts
@@ -1,0 +1,1 @@
+export { ContentList } from "./content-list";
