--- packages/web/src/components/image-block/index.ts
@@ -1,1 +1,0 @@
-export { default as ImageBlock } from "./image-block";
