+++ packages/web/src/components/index.ts
@@ -1,4 +1,5 @@
-export * from "./link-block";
-export * from "./project-block";
-export * from "./project-list";
-export * from "./separator-block";
+export * from "./content-list";
+export * from "./link-list-item";
+export * from "./project-details-list-item";
+export * from "./project-list-item";
+export * from "./separator-list-item";
