--- packages/web/src/components/link-block/index.ts
@@ -1,1 +1,0 @@
-export { default as LinkBlock } from "./link-block";
