--- packages/web/src/components/project-block/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectBlock } from "./project-block";
