--- packages/web/src/components/project-block/project-block.module.css
@@ -1,146 +1,0 @@
-:root {
-  --project-color: transparent;
-}
-
-.item {
-  background-color: var(--details-background-color);
-}
-
-.header {
-  /* padding: 5px 110px; */
-  align-items: center;
-  border-bottom: var(--border-dotted, 1px dotted #929292);
-  display: flex;
-  padding-block: 5px;
-  padding-inline: 10px;
-  position: relative;
-
-  &:hover {
-    color: var(--color-text-default, #000);
-    background-color: var(--project-color, transparent);
-  }
-
-  .active & {
-    color: var(--color-text-default, #000);
-    text-transform: uppercase;
-    font-family: var(--font-bold, "museo_bold", sans-serif);
-    background-color: var(--project-color, transparent);
-  }
-}
-
-.left {
-  flex: 0 0 var(--list-item-left-element-width);
-}
-
-.center {
-  color: currentColor;
-  flex: 1;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  z-index: 1;
-}
-
-.right {
-  flex: 0 0 var(--list-item-right-element-width);
-  text-align: end;
-}
-
-.text {
-  color: var(--color-text-default, #000);
-
-  p {
-    color: var(--color-text-default, #000);
-  }
-}
-
-/* Active state styles */
-/* .active .header {
-  color: var(--color-text-default, #000);
-  text-transform: uppercase;
-  font-family: var(--font-bold, "museo_bold", sans-serif);
-  background-color: var(--project-color, transparent);
-} */
-
-/* .center {
-  text-overflow: ellipsis;
-  overflow: hidden;
-  z-index: 1;
-}
-
-.left {
-  position: absolute;
-  left: 0;
-  top: 5px;
-  padding-left: 10px;
-  z-index: 1;
-}
-
-.right {
-  position: absolute;
-  right: 0;
-  top: 5px;
-  padding-right: 10px;
-  z-index: 1;
-} */
-
-.link {
-  position: absolute;
-  inset: 0;
-  z-index: 50;
-  display: block;
-}
-
-/* Details section (shown when active) */
-.detailsRow {
-  background-color: var(--details-background-color);
-  display: flex;
-  padding: 10px;
-  border-bottom: var(--border-dotted, 1px dotted #929292);
-}
-
-.detailsRowTitle {
-  flex: 0 0 var(--list-item-left-element-width);
-  color: var(--color-text-default, #000);
-}
-
-/* .active .list-item-details {
-  display: none;
-  background-color: var(--details-background-color, red);
-} */
-
-.detailsRowPanelWrapper {
-  flex: 1;
-  position: relative;
-}
-
-.detailsRowPanel {
-  &:not(:first-child) {
-    margin-block-start: 10px;
-  }
-}
-
-/* .paragraph {
-  border-bottom: var(--border-dotted, 1px dotted #929292);
-  padding: 10px 10px 0 10px;
-  position: relative;
-} */
-
-.paragraphTitle {
-  color: var(--color-text-default, #000);
-}
-
-.detailsLeft {
-  position: absolute;
-  width: 90px;
-  left: 10px;
-  top: 10px;
-  min-height: 16px;
-  overflow-wrap: anywhere;
-  word-wrap: break-word;
-  hyphens: auto;
-}
-
-.detailsRight {
-  margin: 0 0 10px 100px;
-  min-height: 16px;
-}
