+++ packages/web/src/components/project-details-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./project-details-list-item";
