+++ packages/web/src/components/project-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./project-list-item";
