--- packages/web/src/components/project-list/index.ts
@@ -1,1 +1,0 @@
-export { ProjectList } from "./project-list";
