+++ packages/web/src/components/project-preview-images/index.ts
@@ -1,0 +1,1 @@
+export * from "./project-preview-images";
