+++ packages/web/src/components/project-preview-images/project-preview-images.module.css
@@ -1,20 +1,10 @@
 .container {
-  /* margin: 0 0 10px 100px; */
   min-height: 16px;
   display: flex;
   flex-wrap: wrap;
   row-gap: 10px;
   column-gap: 8px;
 }
-
-/* .container img {
-  margin: 0 8px 10px 0;
-} */
-
-/* .thumbnailLink {
-  display: inline-block;
-  overflow: hidden;
-} */
 
 .thumbnail {
   display: block;
