--- packages/web/src/components/separator-block/index.ts
@@ -1,1 +1,0 @@
-export { default as SeparatorBlock } from "./separator-block";
