--- packages/web/src/components/separator-block/separator-block.tsx
@@ -1,7 +1,0 @@
-import styles from "./separator-block.module.css";
-
-export default function SeparatorBlock() {
-  return (
-    <li className={styles.separator} />
-  );
-}
