+++ packages/web/src/components/separator-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./separator-list-item";
