--- packages/web/src/components/text-block/index.ts
@@ -1,1 +1,0 @@
-export { default as TextBlock } from "./text-block";
