--- packages/web/src/components/text-block/text-block.module.css
@@ -1,16 +1,0 @@
-.detailsRight {
-  margin: 0 0 10px 100px;
-  min-height: 16px;
-}
-
-.text {
-  color: var(--color-text-default, #000);
-}
-
-.text p {
-  color: var(--color-text-default, #000);
-}
-
-.textpageLink {
-  display: block;
-}
